chore: refactor binary code
