--- src/menu/compute_binary/binary_description.c
@@ -1,11 +1,0 @@
-#include "../../utils/utils.h"
-
-void binary_description(void) {
-  text_blue("--------------------------------------------------------", 15);
-  add_new_line(2);
-  text_green("Binary Conversion", 34);
-  add_new_line(2);
-  text_white("Convert binary values to other system.", 24);
-  add_new_line(2);
-  text_blue("--------------------------------------------------------", 15);
-}
